ecom_site/requirements.txt
apturl==0.5.2
asgiref==3.3.1
asn1crypto==0.24.0
beautifulsoup4==4.6.0
blinker==1.4
Brlapi==0.6.6
cachetools==4.2.0
certifi==2018.1.18
cffi==1.14.4
chardet==3.0.4
checkbox-support==0.22
click==7.1.2
cryptography==3.1
cupshelpers==1.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==3.1.4
django-allauth==0.44.0
django-autoslug==1.9.8
django-braces==1.14.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-friendship==1.9.1
django-gcloud-storage==0.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.1.0
django-social-share==2.1.0
django-stdimage==5.1.1
django-storages==1.7.1
djangorestframework==3.12.2
docutils==0.14
feedparser==5.2.1
flake8==3.5.0
freeze==3.0
google-api-core==1.22.1
google-auth==1.19.2
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
gTTS==2.1.1
gTTS-token==1.1.3
guacamole==0.9.2
gunicorn==19.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
imagesize==1.1.0
Jinja2==2.10
jmespath==0.9.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
onboard==1.4.1
padme==1.1.1
pbr==5.5.1
pexpect==4.2.1
Pillow==7.2.0
plainbox==0.25
progressbar2==3.53.1
protobuf==3.0.0
psycopg2==2.7.7
pyasn1==0.4.2
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyflakes==1.6.0
pygobject==3.26.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-decouple==3.3
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.4
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rsa==4.6
s3transfer==0.2.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
social-auth-core==3.3.3
sqlparse==0.4.1
system-service==0.3
systemd-python==234
testresources==2.0.1
typecheck-decorator==1.3
typing==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
webencodings==0.5
whitenoise==5.2.0
xkit==0.0.0
XlsxWriter==0.9.6
zope.interface==4.3.2
